Refactor as core file
